Moving AI, Player code to respective classes for Update, Move, Render, Reset functions.  Adding TimeStep and GameLoop control for FPS.  Changed Entity base class to use Double for Position to allow for TimeStep usage.  Changed level to lock the AI Spawn zone to eliminate special case checks in movement code.  And general code clean-up.

+++ Pac++Man/CreditsBoard.cpp
@@ -19,8 +19,8 @@ at the bottom of the screen.
 void CreditsBoard::Render() {
     if (isInvalidated) {
         COORD Position;
-        Position.X = xPos;
-        Position.Y = yPos;
+        Position.X = (int)xPos;
+        Position.Y = (int)yPos;
         SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), 12);
         SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), Position);
         std::cout << CREDITS_NAME_TEXT << creditTotal;
