+++ Pac++Man/LivesBoard.cpp
@@ -19,8 +19,8 @@ Comments: Renders the number of Player lives that are currently Active.
 void LivesBoard::Render() {
     if (isInvalidated) {
         COORD Position;
-        Position.X = xPos;
-        Position.Y = yPos;
+        Position.X = (int)xPos;
+        Position.Y = (int)yPos;
         SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), Position);
         std::cout << "\033[40;37;1m" << LIVES_NAME_TEXT;
         Position.Y += 1;
